requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
capx-core==0.1.1
numpy==2.2.6
pillow==12.1.0